chore(docs): diagrams as code with likec4 (#2000)
## Description
We recently decided to use diagrams as code for creating and maintaining
our diagrams. This should reduce maintenance burden and speed up
creation of new diagrams.

Because this is diagrams as code, we lose some ability to customize how
the diagram is generated. There is the ability with likec4 to manually
edit and arrange diagrams, however doing this has some big pitfalls:

1. It causes headaches anytime we update the diagrams because the
generator is not smart enough to take the customizations and add new
things elegantly
2. When exporting the manually created diagrams, the export
functionality is not smart enough to figure out where the center of the
diagram is, resulting in a very large capture space which is mostly just
white space.

Currently created diagrams:
- Overview Diagrams
  - Basic Overview
  - Namespace Overview
  - Ingress / Egress Overview
- Functional Layers (each layer is it's own diagram)
- Namespaces (each namespace is it's own diagram)
- Ports and Protocols

## Open Questions
- How do we like the colors?
- Do we want all these diagrams to be generated and stored in UDS Core
right? Or do we only want a subset of them?
- Does anyone have any experience with likec4 and how to make manually
customized diagrams export the right size? Or some automated way of
trimming whitespace from a png?

## Related Issue

Fixes #1955

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed