fix: prevent network policies from selecting non-authservice SSO clients (#2127)
## Description

This PR fixes a bug where NetPols could be generated for SSO clients
that were not "authservice-enabled" because a missing
`enableAuthserviceSelector` was effectively treated as namespace-wide.

Pulled an existing authservice enabled sso client filter into a new
utility to function that can be used to pre filter for the clients that
we need. Implemented a bunch of new tests for the utility and the
functions that use that utility to make sure i maintained existing
expected behavior. Updated a few different spots that were filtering for
authservice clients in different ways.
 
## Related Issue

Fixes #2120 

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate
- `uds run test:uds-core-e2e`
- `uds run test:unit-tests`

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed