fix: renovate matching for falco rules (#2188)
## Description

Updates the tasks for falco rules to properly match the tags for the
upstream rules. Currently renovate is using a semver-coerced versioning
for these github raw strings that doesn't appear to support/properly
match the new upstream tags.

This change pulls the semver portion out to an env, then uses a renovate
`extractVersion` to match the proper upstream tags. This approach was
tested on my personal repo:
https://github.com/mjnagel/test/pull/122/files

Note that I did not update the tags in this PR, so we should have a
follow-on Renovate PR that actually bumps these tags and the rules.

## Related Issue

Relates to https://github.com/defenseunicorns/uds-core/issues/2185

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)